move search component to views
