--- src/views/HomeView.vue
@@ -1,7 +1,0 @@
-<script setup>
-</script>
-
-<template>
-  <main>
-  </main>
-</template>
